Move baidu provider to addresses folder
